vm: move OpRef to vm::Contracts module
