+++ src/validation/consignment.rs
@@ -31,7 +31,8 @@ use amplify::confinement::Confined;
 use strict_types::TypeSystem;
 
 use super::EAnchor;
-use crate::{BundleId, Genesis, OpId, OpRef, Operation, Schema, TransitionBundle, XWitnessId};
+use crate::vm::OpRef;
+use crate::{BundleId, Genesis, OpId, Operation, Schema, TransitionBundle, XWitnessId};
 
 pub const CONSIGNMENT_MAX_LIBS: usize = 1024;
